Bugfix type errors and add test case for the timer overflowing

+++ cpp/tempenlib/Blinky.h
@@ -13,7 +13,7 @@ public:
 private:
     Stopwatch *stopwatch;
     Led *led;
-    int lastStateChange;
+    unsigned long lastStateChange;
     void flipLedState();
 };
